Moved classes/methods into places where they're used.
