+++ temporal-sdk/src/test/java/io/temporal/workflow/activityTests/AsyncUntyped2ActivityTest.java
@@ -46,8 +46,6 @@ public class AsyncUntyped2ActivityTest {
 
   @Test
   public void testAsyncUntyped2Activity() {
-    // TODO: (vkoby) See if this activityImpl could be constructed from within the rule with the
-    // right completion client.
     activitiesImpl.completionClient =
         testWorkflowRule.getWorkflowClient().newActivityCompletionClient();
     TestWorkflows.TestWorkflow1 client =
