+++ temporal-sdk/src/test/java/io/temporal/workflow/activityTests/AsyncUntypedActivityTest.java
@@ -47,8 +47,6 @@ public class AsyncUntypedActivityTest {
 
   @Test
   public void testAsyncUntypedActivity() {
-    // TODO: (vkoby) See if this activityImpl could be constructed from within the rule with the
-    // right completion client.
     activitiesImpl.completionClient =
         testWorkflowRule.getWorkflowClient().newActivityCompletionClient();
     TestWorkflows.TestWorkflow1 client =
