+++ temporal-sdk/src/test/java/io/temporal/workflow/childWorkflowTests/UntypedChildStubWorkflowTest.java
@@ -19,7 +19,11 @@
 
 package io.temporal.workflow.childWorkflowTests;
 
-import io.temporal.workflow.WorkflowTest;
+import static org.junit.Assert.assertEquals;
+
+import io.temporal.workflow.ChildWorkflowOptions;
+import io.temporal.workflow.ChildWorkflowStub;
+import io.temporal.workflow.Workflow;
 import io.temporal.workflow.shared.SDKTestWorkflowRule;
 import io.temporal.workflow.shared.TestActivities;
 import io.temporal.workflow.shared.TestMultiargdsWorkflowFunctions;
@@ -37,7 +41,7 @@ public class UntypedChildStubWorkflowTest {
   public SDKTestWorkflowRule testWorkflowRule =
       SDKTestWorkflowRule.newBuilder()
           .setWorkflowTypes(
-              WorkflowTest.TestUntypedChildStubWorkflow.class,
+              TestUntypedChildStubWorkflow.class,
               TestMultiargdsWorkflowFunctions.TestMultiargsWorkflowsImpl.class)
           .setActivityImplementations(activitiesImpl)
           .build();
@@ -47,5 +51,58 @@ public class UntypedChildStubWorkflowTest {
     TestWorkflows.TestWorkflow1 client =
         testWorkflowRule.newWorkflowStub200sTimeoutOptions(TestWorkflows.TestWorkflow1.class);
     Assert.assertEquals(null, client.execute(testWorkflowRule.getTaskQueue()));
+  }
+
+  public static class TestUntypedChildStubWorkflow implements TestWorkflows.TestWorkflow1 {
+
+    @Override
+    public String execute(String taskQueue) {
+      ChildWorkflowOptions workflowOptions =
+          ChildWorkflowOptions.newBuilder().setTaskQueue(taskQueue).build();
+      ChildWorkflowStub stubF =
+          Workflow.newUntypedChildWorkflowStub("TestMultiargsWorkflowsFunc", workflowOptions);
+      assertEquals("func", stubF.execute(String.class));
+      // Workflow type overridden through the @WorkflowMethod.name
+      ChildWorkflowStub stubF1 = Workflow.newUntypedChildWorkflowStub("func1", workflowOptions);
+      assertEquals("1", stubF1.execute(String.class, "1"));
+      ChildWorkflowStub stubF2 =
+          Workflow.newUntypedChildWorkflowStub("TestMultiargsWorkflowsFunc2", workflowOptions);
+      assertEquals("12", stubF2.execute(String.class, "1", 2));
+      ChildWorkflowStub stubF3 =
+          Workflow.newUntypedChildWorkflowStub("TestMultiargsWorkflowsFunc3", workflowOptions);
+      assertEquals("123", stubF3.execute(String.class, "1", 2, 3));
+      ChildWorkflowStub stubF4 =
+          Workflow.newUntypedChildWorkflowStub("TestMultiargsWorkflowsFunc4", workflowOptions);
+      assertEquals("1234", stubF4.execute(String.class, "1", 2, 3, 4));
+      ChildWorkflowStub stubF5 =
+          Workflow.newUntypedChildWorkflowStub("TestMultiargsWorkflowsFunc5", workflowOptions);
+      assertEquals("12345", stubF5.execute(String.class, "1", 2, 3, 4, 5));
+      ChildWorkflowStub stubF6 =
+          Workflow.newUntypedChildWorkflowStub("TestMultiargsWorkflowsFunc6", workflowOptions);
+      assertEquals("123456", stubF6.execute(String.class, "1", 2, 3, 4, 5, 6));
+
+      ChildWorkflowStub stubP =
+          Workflow.newUntypedChildWorkflowStub("TestMultiargsWorkflowsProc", workflowOptions);
+      stubP.execute(Void.class);
+      ChildWorkflowStub stubP1 =
+          Workflow.newUntypedChildWorkflowStub("TestMultiargsWorkflowsProc1", workflowOptions);
+      stubP1.execute(Void.class, "1");
+      ChildWorkflowStub stubP2 =
+          Workflow.newUntypedChildWorkflowStub("TestMultiargsWorkflowsProc2", workflowOptions);
+      stubP2.execute(Void.class, "1", 2);
+      ChildWorkflowStub stubP3 =
+          Workflow.newUntypedChildWorkflowStub("TestMultiargsWorkflowsProc3", workflowOptions);
+      stubP3.execute(Void.class, "1", 2, 3);
+      ChildWorkflowStub stubP4 =
+          Workflow.newUntypedChildWorkflowStub("TestMultiargsWorkflowsProc4", workflowOptions);
+      stubP4.execute(Void.class, "1", 2, 3, 4);
+      ChildWorkflowStub stubP5 =
+          Workflow.newUntypedChildWorkflowStub("TestMultiargsWorkflowsProc5", workflowOptions);
+      stubP5.execute(Void.class, "1", 2, 3, 4, 5);
+      ChildWorkflowStub stubP6 =
+          Workflow.newUntypedChildWorkflowStub("TestMultiargsWorkflowsProc6", workflowOptions);
+      stubP6.execute(Void.class, "1", 2, 3, 4, 5, 6);
+      return null;
+    }
   }
 }
